server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
stripe>=11.0.0
python-dotenv>=1.0.0
email-validator>=2.2.0
